chore: Splitting products additional form
